vehicle_receiver/requirements.txt
websockets==12.0
pyserial==3.5
python-dotenv==1.0.1
